feat: Prometheus middleware to support paths to exclude
Currently there is no way to exclude some paths from Prometheus
registry. Sometimes this might be useful, when we want to gather
metrics only for resources that encapsulate some business
related logic.

The proposed changes are about to enable us with such feature.
So, in case you need to exclude some paths, just pass argument
`excluded_paths` with the list of paths and they will not be
shown in metrics report.

Example::
```
app.middlewares.append(
	prometheus_middleware_factory(
		registry=registry,
		excluded_paths=('metrics', 'health', 'something-internal'))
	)
)
```